server/requirements.txt
fastapi==0.104.1
python-socketio==5.10.0
uvicorn==0.24.0
pydantic==2.5.2
numpy==1.26.2
matplotlib==3.8.2
ahrs==0.3.1
filterpy==1.4.5
requests==2.31.0
beautifulsoup4==4.12.2
libsql-client==0.3.1 